Add and install pkg-config files on Windows.
The build system does not provide these as with the Linux/OSX builds,
but nevertheless some downstream dependencies on Windows (e.g. `gtk3`,
`gobject-introspection`) would prefer to use them. The pkg-config files
are created using a patch and envsubst environment variable substitution
to set the package version and prefix path.