recipe/windows-soversion-continuity.patch: add this
This patch results in our DLLs being named like `cairo.dll`, in continuity with
the pre-Meson packages, as opposed to `cairo-2.dll`, which is what the unpatched
Meson build system would produce. This way, our packages will (hopefully) remain
binary compatible across the build system transition.

If we ever find it worthwhile to drop this patch, we could produce a version of
this package with copies of the DLLs under both the versioned and unversioned
names, and development files pointing towards the versioned names. After a round
of rebuilds of dependents, it should be possible to get rid of the unversioned
DLLs.